<==== Mercurial data ====>
Branch: master

+++ client/src/ui/workbench/workbench_navigator.cpp
@@ -730,10 +730,10 @@ void QnWorkbenchNavigator::updateSliderFromReader(bool keepInWindow) {
         startTimeUSec = startTimeMSec * 1000;
     } else {
         endTimeUSec = reader->endTime();
-        endTimeMSec = endTimeUSec == DATETIME_NOW ? qnSyncTime->currentMSecsSinceEpoch() : (endTimeUSec == AV_NOPTS_VALUE ? m_timeSlider->maximum() : endTimeUSec / 1000);
+        endTimeMSec = endTimeUSec == DATETIME_NOW ? qnSyncTime->currentMSecsSinceEpoch() : ((quint64)endTimeUSec == AV_NOPTS_VALUE ? m_timeSlider->maximum() : endTimeUSec / 1000);
 
         startTimeUSec = reader->startTime();                       /* vvvvv  If nothing is recorded, set minimum to end - 10s. */
-        startTimeMSec = startTimeUSec == DATETIME_NOW ? endTimeMSec - 10000 : (startTimeUSec == AV_NOPTS_VALUE ? m_timeSlider->minimum() : startTimeUSec / 1000); 
+        startTimeMSec = startTimeUSec == DATETIME_NOW ? endTimeMSec - 10000 : ((quint64)startTimeUSec == AV_NOPTS_VALUE ? m_timeSlider->minimum() : startTimeUSec / 1000);
     }
 
     m_timeSlider->setRange(startTimeMSec, endTimeMSec);
@@ -742,17 +742,17 @@ void QnWorkbenchNavigator::updateSliderFromReader(bool keepInWindow) {
 
     if(!m_pausedOverride) {
         qint64 timeUSec = m_currentMediaWidget->display()->camDisplay()->isRealTimeSource() ? DATETIME_NOW : m_currentMediaWidget->display()->camera()->getCurrentTime();
-        if(isSearch && timeUSec == AV_NOPTS_VALUE) {
+        if(isSearch && (quint64)timeUSec == AV_NOPTS_VALUE) {
             timeUSec = m_currentMediaWidget->item()->data<qint64>(Qn::ItemTimeRole, AV_NOPTS_VALUE);
-            if(timeUSec != AV_NOPTS_VALUE)
+            if((quint64)timeUSec != AV_NOPTS_VALUE)
                 timeUSec *= 1000;
         }
-        qint64 timeMSec = timeUSec == DATETIME_NOW ? endTimeMSec : (timeUSec == AV_NOPTS_VALUE ? m_timeSlider->value() : timeUSec / 1000);
+        qint64 timeMSec = timeUSec == DATETIME_NOW ? endTimeMSec : ((quint64)timeUSec == AV_NOPTS_VALUE ? m_timeSlider->value() : timeUSec / 1000);
         qint64 timeNext = m_currentMediaWidget->display()->camDisplay()->isRealTimeSource() ? AV_NOPTS_VALUE : m_currentMediaWidget->display()->camDisplay()->getNextTime();
 
-        if (timeUSec != DATETIME_NOW && timeUSec != AV_NOPTS_VALUE){
+        if (timeUSec != DATETIME_NOW && (quint64)timeUSec != AV_NOPTS_VALUE){
             qint64 now = QDateTime::currentDateTimeUtc().toMSecsSinceEpoch();
-            if (m_lastUpdateSlider && m_lastCameraTime == timeMSec && timeNext != AV_NOPTS_VALUE && timeNext - timeMSec <= MAX_FRAME_DURATION){
+            if (m_lastUpdateSlider && m_lastCameraTime == timeMSec && (quint64)timeNext != AV_NOPTS_VALUE && timeNext - timeMSec <= MAX_FRAME_DURATION){
                 qint64 timeDiff = (now - m_lastUpdateSlider) * speed();
                 if (timeDiff < MAX_FRAME_DURATION)
                     timeMSec += timeDiff;
@@ -764,7 +764,7 @@ void QnWorkbenchNavigator::updateSliderFromReader(bool keepInWindow) {
 
         m_timeSlider->setValue(timeMSec, keepInWindow);
 
-        if(timeUSec != AV_NOPTS_VALUE)
+        if((quint64)timeUSec != AV_NOPTS_VALUE)
             updateLive();
 
         if(isSearch) {
@@ -778,7 +778,7 @@ void QnWorkbenchNavigator::updateSliderFromReader(bool keepInWindow) {
         }
     }
 
-    if(!m_currentWidgetLoaded && (startTimeUSec != AV_NOPTS_VALUE && endTimeUSec != AV_NOPTS_VALUE)) {
+    if(!m_currentWidgetLoaded && ((quint64)startTimeUSec != AV_NOPTS_VALUE && (quint64)endTimeUSec != AV_NOPTS_VALUE)) {
         m_currentWidgetLoaded = true;
         updateTargetPeriod();
